+++ src/main/java/com/caronte/caronte/_Customer/CustomerRepository.java
@@ -1,4 +1,4 @@
-package com.caronte.caronte.customer;
+package com.caronte.caronte._Customer;
 
 
 import java.util.Optional;
