+++ src/main/java/com/caronte/caronte/_Plan/Plan.java
@@ -1,4 +1,4 @@
-package com.caronte.caronte.plan;
+package com.caronte.caronte._Plan;
 
 import java.time.LocalDate;
