+++ src/main/java/com/caronte/caronte/_User/UserRepository.java
@@ -1,4 +1,4 @@
-package com.caronte.caronte.user;
+package com.caronte.caronte._User;
 
 import java.util.Optional;
